+++ index.go
@@ -11,22 +11,30 @@ import (
 	"net/http"
 	"sort"
 	"strconv"
+	"strings"
 	"time"
 )
 
 type IndexPageData struct {
-	Version        string
-	Company        string
-	Alarms         string
-	MenuOverview   string
-	MenuWorkplaces string
-	MenuCharts     string
-	MenuStatistics string
-	MenuData       string
-	MenuSettings   string
-	Compacted      string
-	UserEmail      string
-	UserName       string
+	Version               string
+	Company               string
+	Alarms                string
+	MenuOverview          string
+	MenuWorkplaces        string
+	MenuCharts            string
+	MenuStatistics        string
+	MenuData              string
+	MenuSettings          string
+	Compacted             string
+	UserEmail             string
+	UserName              string
+	Workplaces            []IndexWorkplaceSelection
+	DataFilterPlaceholder string
+}
+
+type IndexWorkplaceSelection struct {
+	WorkplaceName      string
+	WorkplaceSelection string
 }
 
 type IndexData struct {
@@ -77,11 +85,11 @@ func loadIndexData(writer http.ResponseWriter, request *http.Request, _ httprout
 	}
 	data := IndexData{}
 	loc, err := time.LoadLocation(location)
-	workplaceNames, workplacePercents := downloadProductionData(db, loc)
-	terminalDowntimeNames, terminalDowntimeValues := downloadTerminalDowntimeData(db)
-	terminalBreakdownNames, terminalBreakdownValues := downloadTerminalBreakdownData(db)
-	terminalAlarmNames, terminalAlarmValues := downloadTerminalAlarmData(db)
-	calendarData := downloadCalendarData(db, loc)
+	workplaceNames, workplacePercents := downloadProductionData(db, loc, email)
+	terminalDowntimeNames, terminalDowntimeValues := downloadTerminalDowntimeData(db, email)
+	terminalBreakdownNames, terminalBreakdownValues := downloadTerminalBreakdownData(db, email)
+	terminalAlarmNames, terminalAlarmValues := downloadTerminalAlarmData(db, email)
+	calendarData := downloadCalendarData(db, loc, email)
 	data.WorkplaceNames = workplaceNames
 	data.WorkplacePercents = workplacePercents
 	data.TerminalDowntimeNames = terminalDowntimeNames
@@ -109,30 +117,50 @@ func loadIndexData(writer http.ResponseWriter, request *http.Request, _ httprout
 	logInfo("INDEX", "Index data sent in "+time.Since(timer).String())
 }
 
-func downloadCalendarData(db *gorm.DB, loc *time.Location) [][]string {
-	stateRecordsAsMap := downloadData(db, time.Now().In(loc).AddDate(-1, 0, 0), time.Now().In(loc), 0, loc)
+func downloadCalendarData(db *gorm.DB, loc *time.Location, email string) [][]string {
+	stateRecordsAsMap := downloadData(db, time.Now().In(loc).AddDate(-1, 0, 0), time.Now().In(loc), 0, loc, email)
 	var calendarData [][]string
 	for key, value := range stateRecordsAsMap {
 		if key != time.Now().In(loc).Format("2006-01-02") {
 			totalDayDuration := 24 * time.Hour
-			percentage := strconv.FormatFloat(value.Seconds()/(totalDayDuration.Seconds()*float64(len(cachedWorkplacesById)))*100, 'f', 1, 64)
+			percentage := ""
+			if len(cachedUserSettings[email].selectedWorkplaces) == 0 {
+				percentage = strconv.FormatFloat(value.Seconds()/(totalDayDuration.Seconds()*float64(len(cachedWorkplacesById)))*100, 'f', 1, 64)
+			} else {
+				percentage = strconv.FormatFloat(value.Seconds()/(totalDayDuration.Seconds()*float64(len(cachedUserSettings[email].selectedWorkplaces)))*100, 'f', 1, 64)
+			}
 			calendarData = append(calendarData, []string{key, percentage})
 		} else {
 			startOfToday := time.Date(time.Now().In(loc).Year(), time.Now().In(loc).Month(), time.Now().In(loc).Day(), 0, 0, 0, 0, loc)
 			totalTodayDuration := time.Now().In(loc).Sub(startOfToday)
-			percentage := strconv.FormatFloat(value.Seconds()/(totalTodayDuration.Seconds()*float64(len(cachedWorkplacesById)))*100, 'f', 1, 64)
+			percentage := ""
+			if len(cachedUserSettings[email].selectedWorkplaces) == 0 {
+				percentage = strconv.FormatFloat(value.Seconds()/(totalTodayDuration.Seconds()*float64(len(cachedWorkplacesById)))*100, 'f', 1, 64)
+			} else {
+				percentage = strconv.FormatFloat(value.Seconds()/(totalTodayDuration.Seconds()*float64(len(cachedUserSettings[email].selectedWorkplaces)))*100, 'f', 1, 64)
+			}
 			calendarData = append(calendarData, []string{key, percentage})
 		}
 	}
 	return calendarData
 }
 
-func downloadData(db *gorm.DB, fromDate time.Time, toDate time.Time, workplaceId uint, loc *time.Location) map[string]time.Duration {
+func downloadData(db *gorm.DB, fromDate time.Time, toDate time.Time, workplaceId uint, loc *time.Location, email string) map[string]time.Duration {
 	stateRecordsAsMap := make(map[string]time.Duration)
 	var stateRecords []database.StateRecord
 	if workplaceId == 0 {
-		for _, workplace := range cachedWorkplacesById {
-			db.Select("state_id, date_time_start").Where("date_time_start >= ?", fromDate).Where("date_time_start <= ?", toDate).Where("workplace_id = ?", workplace.ID).Order("date_time_start asc").Find(&stateRecords)
+		var workplaces []database.Workplace
+		if len(cachedUserSettings[email].selectedWorkplaces) == 0 {
+			for _, workplace := range cachedWorkplacesByName {
+				workplaces = append(workplaces, workplace)
+			}
+		} else {
+			for _, workplace := range cachedUserSettings[email].selectedWorkplaces {
+				workplaces = append(workplaces, cachedWorkplacesByName[workplace])
+			}
+		}
+		for _, workplace := range workplaces {
+			db.Select("state_id, date_time_start").Where("date_time_start >= ?", fromDate).Where("date_time_start <= ?", toDate).Where("workplace_id = ?", workplace.ID).Order("id asc").Find(&stateRecords)
 			for index, record := range stateRecords {
 				nextDate := time.Now().In(loc)
 				if index < len(stateRecords)-1 {
@@ -142,7 +170,6 @@ func downloadData(db *gorm.DB, fromDate time.Time, toDate time.Time, workplaceId
 					if record.DateTimeStart.In(loc).Day() == nextDate.In(loc).Day() {
 						stateRecordsAsMap[record.DateTimeStart.In(loc).Format("2006-01-02")] += nextDate.In(loc).Sub(record.DateTimeStart.In(loc))
 					} else {
-
 						endOfRecordDay := time.Date(record.DateTimeStart.In(loc).Year(), record.DateTimeStart.In(loc).Month(), record.DateTimeStart.In(loc).Day()+1, 0, 0, 0, 0, loc)
 						for record.DateTimeStart.In(loc).Before(nextDate.In(loc)) {
 							stateRecordsAsMap[record.DateTimeStart.In(loc).Format("2006-01-02")] += endOfRecordDay.In(loc).Sub(record.DateTimeStart.In(loc))
@@ -157,7 +184,7 @@ func downloadData(db *gorm.DB, fromDate time.Time, toDate time.Time, workplaceId
 		}
 
 	} else {
-		db.Select("state_id, date_time_start").Where("date_time_start >= ?", fromDate).Where("date_time_start <= ?", toDate).Where("workplace_id = ?", workplaceId).Order("date_time_start asc").Find(&stateRecords)
+		db.Select("state_id, date_time_start").Where("date_time_start >= ?", fromDate).Where("date_time_start <= ?", toDate).Where("workplace_id = ?", workplaceId).Order("id asc").Find(&stateRecords)
 		for index, record := range stateRecords {
 			nextDate := time.Now().In(loc)
 			if index < len(stateRecords)-1 {
@@ -183,9 +210,19 @@ func downloadData(db *gorm.DB, fromDate time.Time, toDate time.Time, workplaceId
 	return stateRecordsAsMap
 }
 
-func downloadTerminalAlarmData(db *gorm.DB) ([]string, []float64) {
+func downloadTerminalAlarmData(db *gorm.DB, email string) ([]string, []float64) {
 	var alarmRecords []database.AlarmRecord
-	db.Where("date_time_end is null").Find(&alarmRecords)
+	if len(cachedUserSettings[email].selectedWorkplaces) > 0 {
+		workplaceIds := `workplace_id in ('`
+		for _, workplace := range cachedUserSettings[email].selectedWorkplaces {
+			workplaceIds += strconv.Itoa(int(cachedWorkplacesByName[workplace].ID)) + `','`
+		}
+		workplaceIds = strings.TrimSuffix(workplaceIds, `,'`)
+		workplaceIds += ")"
+		db.Where("date_time_end is null").Where(workplaceIds).Find(&alarmRecords)
+	} else {
+		db.Where("date_time_end is null").Find(&alarmRecords)
+	}
 	var indexDataWorkplaces []IndexDataWorkplace
 	for _, alarmRecord := range alarmRecords {
 		var indexDataWorkplace IndexDataWorkplace
@@ -205,10 +242,19 @@ func downloadTerminalAlarmData(db *gorm.DB) ([]string, []float64) {
 	return terminalAlarmNames, terminalAlarmValues
 }
 
-func downloadTerminalBreakdownData(db *gorm.DB) ([]string, []float64) {
+func downloadTerminalBreakdownData(db *gorm.DB, email string) ([]string, []float64) {
 	var breakdownRecords []database.BreakdownRecord
-	db.Where("date_time_end is null").Find(&breakdownRecords)
-
+	if len(cachedUserSettings[email].selectedWorkplaces) > 0 {
+		workplaceIds := `workplace_id in ('`
+		for _, workplace := range cachedUserSettings[email].selectedWorkplaces {
+			workplaceIds += strconv.Itoa(int(cachedWorkplacesByName[workplace].ID)) + `','`
+		}
+		workplaceIds = strings.TrimSuffix(workplaceIds, `,'`)
+		workplaceIds += ")"
+		db.Where("date_time_end is null").Where(workplaceIds).Find(&breakdownRecords)
+	} else {
+		db.Where("date_time_end is null").Find(&breakdownRecords)
+	}
 	var indexDataWorkplaces []IndexDataWorkplace
 	for _, breakdownRecord := range breakdownRecords {
 		var indexDataWorkplace IndexDataWorkplace
@@ -228,9 +274,19 @@ func downloadTerminalBreakdownData(db *gorm.DB) ([]string, []float64) {
 	return terminalBreakdownNames, terminalBreakdownValues
 }
 
-func downloadTerminalDowntimeData(db *gorm.DB) ([]string, []float64) {
+func downloadTerminalDowntimeData(db *gorm.DB, email string) ([]string, []float64) {
 	var downtimeRecords []database.DowntimeRecord
-	db.Where("date_time_end is null").Find(&downtimeRecords)
+	if len(cachedUserSettings[email].selectedWorkplaces) > 0 {
+		workplaceIds := `workplace_id in ('`
+		for _, workplace := range cachedUserSettings[email].selectedWorkplaces {
+			workplaceIds += strconv.Itoa(int(cachedWorkplacesByName[workplace].ID)) + `','`
+		}
+		workplaceIds = strings.TrimSuffix(workplaceIds, `,'`)
+		workplaceIds += ")"
+		db.Where("date_time_end is null").Where(workplaceIds).Find(&downtimeRecords)
+	} else {
+		db.Where("date_time_end is null").Find(&downtimeRecords)
+	}
 
 	var indexDataWorkplaces []IndexDataWorkplace
 	for _, downtimeRecord := range downtimeRecords {
@@ -251,12 +307,22 @@ func downloadTerminalDowntimeData(db *gorm.DB) ([]string, []float64) {
 	return terminalDowntimeNames, terminalDowntimeValues
 }
 
-func downloadProductionData(db *gorm.DB, loc *time.Location) ([]string, []float64) {
+func downloadProductionData(db *gorm.DB, loc *time.Location, email string) ([]string, []float64) {
 	var workplaceNames []string
 	var workplacePercents []float64
 	var indexDataWorkplaces []IndexDataWorkplace
-	for _, workplace := range cachedWorkplacesByName {
-		data := downloadData(db, time.Date(time.Now().UTC().Year(), time.Now().UTC().Month(), time.Now().UTC().Day(), 0, 0, 0, 0, time.Now().Location()), time.Now().In(loc), workplace.ID, loc)
+	var workplaces []database.Workplace
+	if len(cachedUserSettings[email].selectedWorkplaces) == 0 {
+		for _, workplace := range cachedWorkplacesByName {
+			workplaces = append(workplaces, workplace)
+		}
+	} else {
+		for _, workplace := range cachedUserSettings[email].selectedWorkplaces {
+			workplaces = append(workplaces, cachedWorkplacesByName[workplace])
+		}
+	}
+	for _, workplace := range workplaces {
+		data := downloadData(db, time.Date(time.Now().UTC().Year(), time.Now().UTC().Month(), time.Now().UTC().Day(), 0, 0, 0, 0, time.Now().Location()), time.Now().In(loc), workplace.ID, loc, email)
 		var totalDuration time.Duration
 		for _, duration := range data {
 			totalDuration = duration
@@ -295,7 +361,18 @@ func index(writer http.ResponseWriter, request *http.Request, _ httprouter.Param
 	data.Compacted = cachedUserSettings[email].menuState
 	data.UserEmail = email
 	data.UserName = cachedUsersByEmail[email].FirstName + " " + cachedUsersByEmail[email].SecondName
-
+	var dataWorkplaces []IndexWorkplaceSelection
+	for _, workplace := range cachedWorkplacesById {
+		dataWorkplaces = append(dataWorkplaces, IndexWorkplaceSelection{
+			WorkplaceName:      workplace.Name,
+			WorkplaceSelection: getWorkplaceSelection(cachedUserSettings[email].selectedWorkplaces, workplace.Name),
+		})
+	}
+	sort.Slice(dataWorkplaces, func(i, j int) bool {
+		return dataWorkplaces[i].WorkplaceName < dataWorkplaces[j].WorkplaceName
+	})
+	data.Workplaces = dataWorkplaces
+	data.DataFilterPlaceholder = getLocale(email, "data-table-search-title")
 	tmpl := template.Must(template.ParseFiles("./html/index.html"))
 	_ = tmpl.Execute(writer, data)
 	logInfo("INDEX", "Index page sent in "+time.Since(timer).String())