+++ main.go
@@ -8,7 +8,7 @@ import (
 	"os"
 )
 
-const version = "2021.2.1.1"
+const version = "2021.2.1.13"
 const serviceName = "System WebService"
 const serviceDescription = "System web interface"
 const config = "user=postgres password=pj79.. dbname=system host=localhost port=5432 sslmode=disable application_name=system_webservice"
@@ -62,6 +62,7 @@ func (p *program) run() {
 	router.GET("/data", basicAuth(data))
 	router.GET("/settings", basicAuth(settings))
 	router.POST("/update_user_settings", updateUserSettings)
+	router.POST("/update_user_workplaces", updateUserWorkplaces)
 	router.POST("/load_index_data", loadIndexData)
 	router.POST("/load_table_data", loadTableData)
 	router.POST("/load_chart_data", loadChartData)